# api-docs.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8180/api/findev","description":"Generated server url"}],"paths":{"/userapi/{user_id}/{project_id}":{"put":{"tags":["application"],"operationId":"removeProject","parameters":[{"name":"user_id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"project_id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserSo"}}}}}},"post":{"tags":["application"],"operationId":"setProject","parameters":[{"name":"user_id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"project_id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserSo"}}}}}}},"/userapi/":{"get":{"tags":["application"],"operationId":"getUsers","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserSo"}}}}}}},"post":{"tags":["application"],"operationId":"createUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserInput"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserSo"}}}}}}},"/projectapi/{id}":{"get":{"tags":["application"],"operationId":"getProjectById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProjectSo"}}}}}},"post":{"tags":["application"],"operationId":"createProject","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProjectInput"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProjectSo"}}}}}},"delete":{"tags":["application"],"operationId":"deleteProject","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"integer","format":"int64"}}}}}}},"/userapi/{id}":{"get":{"tags":["application"],"operationId":"getUserById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserSo"}}}}}},"delete":{"tags":["application"],"operationId":"deleteUser","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"integer","format":"int64"}}}}}}},"/projectapi/":{"get":{"tags":["application"],"operationId":"getProjects","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ProjectSo"}}}}}}}},"/":{"get":{"tags":["application"],"operationId":"getInfo","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}}},"components":{"schemas":{"ProjectEntity":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"title":{"type":"string"},"description":{"type":"string"}},"nullable":true},"UserSo":{"required":["name","surname"],"type":"object","properties":{"name":{"type":"string"},"surname":{"type":"string"},"projects":{"uniqueItems":true,"type":"array","nullable":true,"items":{"$ref":"#/components/schemas/ProjectEntity"}},"id":{"type":"integer","format":"int64"}}},"UserInput":{"required":["name","surname"],"type":"object","properties":{"name":{"type":"string"},"surname":{"type":"string"},"projects":{"uniqueItems":true,"type":"array","nullable":true,"items":{"$ref":"#/components/schemas/ProjectEntity"}}}},"ProjectInput":{"required":["description","title"],"type":"object","properties":{"title":{"type":"string"},"description":{"type":"string"},"users":{"uniqueItems":true,"type":"array","items":{"$ref":"#/components/schemas/UserEntity"}}}},"UserEntity":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"surname":{"type":"string"}}},"ProjectSo":{"required":["description","title"],"type":"object","properties":{"title":{"type":"string"},"description":{"type":"string"},"id":{"type":"integer","format":"int64"},"users":{"uniqueItems":true,"type":"array","items":{"$ref":"#/components/schemas/UserEntity"}}}}}}}